@@ -4,10 +4,8 @@ import com.example.buensabor.entities.comprobantes.Factura;
 import com.example.buensabor.services.comprobantes.FacturaService;
 import com.itextpdf.text.DocumentException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-
-import javax.mail.MessagingException;
+import jakarta.mail.MessagingException;
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
