@@ -1,13 +1,13 @@
 package com.example.buensabor.repositories.articulos;
 
 import com.example.buensabor.entities.articulos.Inventario;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface IInventarioRepository extends IBaseRepository<Inventario,Long> {
+public interface InventarioRepository extends BaseRepository<Inventario,Long> {
 
     List<Inventario> findInventarioByArticuloId(Long id);
 }
