@@ -1,7 +1,7 @@
 package com.example.buensabor.repositories.articulos;
 
 import com.example.buensabor.entities.articulos.RecetaElaborado;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IRecetaElaboradoRepository extends IBaseRepository<RecetaElaborado,Long> {
+public interface RecetaElaboradoRepository extends BaseRepository<RecetaElaborado,Long> {
 
     @Query(value = "SELECT * FROM receta_elaborado r WHERE r.id_articulo_elaborado_detalle = :idArticuloDetalle and r.fecha_baja is null", nativeQuery = true)
     List<RecetaElaborado> getRecetaByIdAed(@Param("idArticuloDetalle") Long idArticuloDetalle);
