@@ -1,9 +1,9 @@
 package com.example.buensabor.repositories.articulos;
 
 import com.example.buensabor.entities.articulos.TipoArticulo;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITipoArticuloRepository extends IBaseRepository<TipoArticulo,Long> {
+public interface TipoArticuloRepository extends BaseRepository<TipoArticulo,Long> {
 }
