@@ -1,7 +1,7 @@
 package com.example.buensabor.repositories.articulos;
 
 import com.example.buensabor.entities.articulos.UnidadMedida;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IUnidadMedidaRepository extends IBaseRepository<UnidadMedida,Long> {
+public interface UnidadMedidaRepository extends BaseRepository<UnidadMedida,Long> {
 
     @Query(value = "SELECT u.* \n" +
                    "FROM receta_elaborado r \n" +
