@@ -1,9 +1,9 @@
 package com.example.buensabor.repositories.comprobantes;
 
 import com.example.buensabor.entities.comprobantes.PedidoEstado;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPedidoEstadoRepository extends IBaseRepository<PedidoEstado, Long> {
+public interface PedidoEstadoRepository extends BaseRepository<PedidoEstado, Long> {
 }
