@@ -1,9 +1,9 @@
 package com.example.buensabor.repositories.comprobantes;
 
 import com.example.buensabor.entities.comprobantes.TipoEnvio;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITipoEnvioRepository extends IBaseRepository<TipoEnvio,Long> {
+public interface TipoEnvioRepository extends BaseRepository<TipoEnvio,Long> {
 }
