@@ -1,7 +1,7 @@
 package com.example.buensabor.repositories.domicilio;
 
 import com.example.buensabor.entities.domicilio.Domicilio;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IDomicilioRepository extends IBaseRepository<Domicilio,Long> {
+public interface DomicilioRepository extends BaseRepository<Domicilio,Long> {
     List<Domicilio> findByFechaBajaIsNull();
 
     @Query(value = "SELECT d FROM Domicilio d WHERE d.usuario.id = :idUsuario AND d.fechaBaja IS null")
