@@ -1,14 +1,14 @@
 package com.example.buensabor.repositories.rubro;
 
 import com.example.buensabor.entities.rubro.Rubro;
-import com.example.buensabor.repositories.base.IBaseRepository;
+import com.example.buensabor.repositories.base.BaseRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface IRubroRepository extends IBaseRepository<Rubro, Long> {
+public interface RubroRepository extends BaseRepository<Rubro, Long> {
 
     List<Rubro> findByFechaBajaIsNull();
 
