@@ -6,12 +6,17 @@ import com.nimbusds.jwt.JWT;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.JWTParser;
 import io.jsonwebtoken.*;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import io.jsonwebtoken.security.SignatureException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Component;
+
+import java.security.Key;
 import java.text.ParseException;
 import java.util.Date;
 import java.util.List;
@@ -37,15 +42,15 @@ public class JwtProvider {
                 .claim("roles", roles)
                 .setIssuedAt(new Date())
                 .setExpiration(new Date(new Date().getTime() + expiration * 180))
-                .signWith(SignatureAlgorithm.HS512, secret.getBytes())
+                .signWith(getSecret(secret))
                 .compact();
     }
     public String getEmailFromToken(String token) {
-        return Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(token).getBody().getSubject();
+        return Jwts.parserBuilder().setSigningKey(getSecret(secret)).build().parseClaimsJws(token).getBody().getSubject();
     }
     public boolean validarToken(String token) {
         try {
-            Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(token);
+            Jwts.parserBuilder().setSigningKey(getSecret(secret)).build().parseClaimsJws(token);
             return true;
         } catch (MalformedJwtException e) {
             logger.error("token mal formado");
@@ -63,7 +68,7 @@ public class JwtProvider {
 
     public String refreshToken(JwtDto jwtDto) throws ParseException {
         try{
-            Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(jwtDto.getToken());
+            Jwts.parserBuilder().setSigningKey(getSecret(secret)).build().parseClaimsJws(jwtDto.getToken());
         }catch (ExpiredJwtException ex){
             JWT jwt = JWTParser.parse(jwtDto.getToken());
             JWTClaimsSet claimsSet = jwt.getJWTClaimsSet();
@@ -74,9 +79,14 @@ public class JwtProvider {
                     .claim("roles", roles)
                     .setIssuedAt(new Date())
                     .setExpiration(new Date(new Date().getTime() + expiration))
-                    .signWith(SignatureAlgorithm.HS512, secret.getBytes())
+                    .signWith(getSecret(secret))
                     .compact();
         }
         return null;
+    }
+
+    private Key getSecret(String secret){
+        byte [] secretBytes = Decoders.BASE64URL.decode(secret);
+        return Keys.hmacShaKeyFor(secretBytes);
     }
 }
