@@ -2,6 +2,7 @@ package com.example.buensabor.security.services;
 
 import com.example.buensabor.security.entities.Usuario;
 import com.example.buensabor.security.entities.UsuarioPrincipal;
+import com.example.buensabor.security.repositories.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -11,16 +12,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
 
-    private final UsuarioService usuarioService;
+    private final UsuarioRepository usuarioRepository;
 
     @Autowired
-    public UserDetailsServiceImpl(UsuarioService usuarioService) {
-        this.usuarioService = usuarioService;
+    public UserDetailsServiceImpl(UsuarioRepository usuarioRepository) {
+        this.usuarioRepository = usuarioRepository;
     }
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Usuario usuario = usuarioService.getByEmail(email).get();
+        Usuario usuario = usuarioRepository.findByEmail(email).get();
         return UsuarioPrincipal.build(usuario);
     }
 }
