@@ -1,7 +1,7 @@
 package com.example.buensabor.services.articulos;
 
 import com.example.buensabor.entities.articulos.ArticuloElaboradoDetalle;
-import com.example.buensabor.repositories.articulos.IArticuloElaboradoDetalleRepository;
+import com.example.buensabor.repositories.articulos.ArticuloElaboradoDetalleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +12,7 @@ import java.util.Optional;
 public class ArticuloElaboradoDetalleService {
 
     @Autowired
-    IArticuloElaboradoDetalleRepository articuloElaboradoDetalleRepository;
+    ArticuloElaboradoDetalleRepository articuloElaboradoDetalleRepository;
 
     public List<ArticuloElaboradoDetalle> getArticuloDetalles() {
         return (List<ArticuloElaboradoDetalle>) articuloElaboradoDetalleRepository.findAll();
