@@ -1,7 +1,7 @@
 package com.example.buensabor.services.articulos;
 
 import com.example.buensabor.entities.articulos.HistoricoArticulo;
-import com.example.buensabor.repositories.articulos.IHistoricoArticuloRepository;
+import com.example.buensabor.repositories.articulos.HistoricoArticuloRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -11,7 +11,7 @@ import java.util.Optional;
 public class HistoricoArticuloService {
 
     @Autowired
-    IHistoricoArticuloRepository historicoArticuloRepository;
+    HistoricoArticuloRepository historicoArticuloRepository;
 
     public List<HistoricoArticulo> getHistoricoArticulos() {
         return (List<HistoricoArticulo>) historicoArticuloRepository.findAll();
