@@ -1,7 +1,7 @@
 package com.example.buensabor.services.comprobantes;
 
 import com.example.buensabor.entities.comprobantes.TipoEnvio;
-import com.example.buensabor.repositories.comprobantes.ITipoEnvioRepository;
+import com.example.buensabor.repositories.comprobantes.TipoEnvioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ import java.util.Optional;
 public class TipoEnvioService {
 
     @Autowired
-    ITipoEnvioRepository tipoEnvioRepository;
+    TipoEnvioRepository tipoEnvioRepository;
 
     public List<TipoEnvio> getTiposEnvio() {
         return (ArrayList<TipoEnvio>) tipoEnvioRepository.findAll();
