@@ -1,7 +1,7 @@
 package com.example.buensabor.services.domicilio;
 
 import com.example.buensabor.entities.domicilio.Localidad;
-import com.example.buensabor.repositories.domicilio.ILocalidadRepository;
+import com.example.buensabor.repositories.domicilio.LocalidadRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +12,7 @@ import java.util.Optional;
 public class LocalidadService {
 
     @Autowired
-    ILocalidadRepository localidadRepository;
+    LocalidadRepository localidadRepository;
 
     public List<Localidad> getLocalidades(){
         return (List<Localidad>) localidadRepository.findAll();
